refactor: WIP
